4. Django/Django/7. django_json_single_model/requirements.txt
asgiref==3.5.0
Django==3.2
django-extensions==3.1.5
django-seed==0.2.0
djangorestframework==3.13.1
Faker==13.3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toposort==1.7
